+++ packages/frontend/src/scripts/embed.ts
@@ -1,7 +1,8 @@
-import {GearPlanSheet, GearPlanSheetGui} from "./components";
 import {LoadingBlocker} from "./components/loader";
 
 import {setTitle} from "./base_ui";
+import {GearPlanSheet} from "./sheet";
+import {GearPlanSheetGui} from "./components/sheet";
 
 let embedDiv: HTMLDivElement;
